requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyMuPDF
python-docx
requests
urllib3
